requirements.txt
syftbox>=0.1.5
python-dotenv==1.0.1
pandas
rapidfuzz
scikit-learn
selenium
diffprivlib>=0.6.5
phe
matplotlib